hive: add optimism devnet simulator (#1)
A new simulator for optimism devnet is added. It requires
multiple optimism specific clients and their sequence is
significant. So the only way to invoke the devnet simulator
is to launch with the following clients:

hive --sim optimism  --client=ops-l1,ops-l2,ops-opnode,ops-l2os --docker.output

The current test suite is adapter from ethereum/rpc and modified to suit
the optimism devnet. Only a subset of the RPC tests are supported at
the moment but more will be added later. Other test suites may extend
the devnet in the future to extend test coverage.